Makes the pre-step more generic
Using language or framework specific labels for the prefix/pre-step is somewhat retrictive and means that a new release of `rind` is neessary for each one added. Not ideal. As such the labels value has now been changed to *be* the pre-step.

e.g.:

`app.rind: python` needs to be changed to `app.rind: source /venv/bin/activate`.

This is combined (if present) to the actual passed in command with an `&&`